+++ inform/__init__.py
@@ -1,5 +1,5 @@
-__version__ = '1.17.2'
-__released__ = '2019-06-11'
+__version__ = '1.17.3'
+__released__ = '2019-06-15'
 
 from .inform import (
     # inform utility functions and classes